Trucate download file name in the middle
