+++ Sources/App/WebView/DownloadManager/ActivityViewController.swift
@@ -1,0 +1,16 @@
+import Foundation
+import SwiftUI
+
+struct ShareWrapper: Identifiable {
+    let id = UUID()
+    let url: URL
+}
+
+struct ActivityViewController: UIViewControllerRepresentable {
+    let shareWrapper: ShareWrapper
+    func makeUIViewController(context: Context) -> UIActivityViewController {
+        UIActivityViewController(activityItems: [shareWrapper.url], applicationActivities: nil)
+    }
+
+    func updateUIViewController(_ uiViewController: UIActivityViewController, context: Context) {}
+}
